+++ src/app/locales/en/components/-storage/xrootd.js
@@ -4,11 +4,11 @@ export default {
     tip: 'Full URL of the XRootD server, including scheme (root or http) and path, e.g. "root://192.168.0.1//data". Please note, that XRootD URL format requires double slash after host to indicate absolute path.',
   },
   fileModeMask: {
-    name: 'File mode mask',
+    name: 'Imported file mode mask',
     tip: 'Defines the file permissions mask, which is used to map XRootD file mode to POSIX mode. For instance a mask "0664" for readable file on XRootD would result in a file which is readable for all users, but file which is writeable in XRootD will be only writeable by user and group.',
   },
   dirModeMask: {
-    name: 'Directory mode mask',
+    name: 'Imported directory mode mask',
     tip: 'Defines the directory permissions mask, which is used to map XRootD dir mode to POSIX mode. For instance a mask "0770" for readable directory on XRootD would result in a directory which is readable for owner and group but not for others.',
   },
   credentialsType: {